Package for correct GUI

+++ utils.py
@@ -75,16 +75,16 @@ def fft(x, frame_width, channels):
     if FRAMEW != frame_width:
         FRAMEW = frame_width
         CHANNELS = channels
-        WINDOW = np.stack([np.hamming(frame_width) ** 0.5] * channels).T
-    return np.fft.rfft(WINDOW * x, axis=0)
+        WINDOW = np.stack([np.hamming(frame_width).astype(np.float32) ** 0.5] * channels).T
+    return np.fft.rfft(WINDOW * x, axis=0).astype(np.complex64)
 
 def ifft(x, frame_width, channels):
     out = np.zeros((frame_width, channels), dtype=np.float32)
     if channels == 1:
-        out =  np.hamming(frame_width)**0.5 *np.fft.irfft(x)
+        out =  np.hamming(frame_width).astype(np.float32) ** 0.5 *np.fft.irfft(x)
     else:
         for ch in range(channels):
-            out[:, ch] = np.fft.irfft(x[:, ch]) * np.hamming(frame_width) ** 0.5
+            out[:, ch] = np.fft.irfft(x[:, ch]).astype(np.float32) * np.hamming(frame_width).astype(np.float32) ** 0.5
     return out
